+++ src/core/include/cesium/omniverse/FabricMaterialDefinition.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "cesium/omniverse/GltfUtil.h"
+#include "cesium/omniverse/MetadataUtil.h"
 
 #include <glm/glm.hpp>
 #include <pxr/base/gf/vec2f.h>
@@ -26,9 +27,7 @@ class FabricMaterialDefinition {
     [[nodiscard]] uint64_t getImageryLayerCount() const;
     [[nodiscard]] bool hasTilesetMaterial() const;
     [[nodiscard]] const pxr::SdfPath& getTilesetMaterialPath() const;
-    [[nodiscard]] const std::vector<MdlInternalPropertyType>& getMdlInternalPropertyAttributePropertyTypes() const;
-    [[nodiscard]] const std::vector<MdlInternalPropertyType>& getMdlInternalPropertyTexturePropertyTypes() const;
-    [[nodiscard]] const std::vector<MdlInternalPropertyType>& getMdlInternalPropertyTablePropertyTypes() const;
+    [[nodiscard]] const std::vector<MetadataUtil::PropertyDefinition>& getProperties() const;
 
     // Make sure to update this function when adding new fields to the class
     bool operator==(const FabricMaterialDefinition& other) const;
@@ -39,9 +38,7 @@ class FabricMaterialDefinition {
     std::vector<FeatureIdType> _featureIdTypes;
     uint64_t _imageryLayerCount;
     pxr::SdfPath _tilesetMaterialPath;
-    std::vector<MdlInternalPropertyType> _mdlInternalPropertyAttributePropertyTypes;
-    std::vector<MdlInternalPropertyType> _mdlInternalPropertyTexturePropertyTypes;
-    std::vector<MdlInternalPropertyType> _mdlInternalPropertyTablePropertyTypes;
+    std::vector<MetadataUtil::PropertyDefinition> _properties;
 };
 
 } // namespace cesium::omniverse